+++ tests/external_data_source/provider.tf
@@ -18,5 +18,5 @@ provider "azurerm" {
 
 # REMEMBER to set AZUREIPAM_API_URL and AZUREIPAM_TOKEN env variables
 provider "azureipam" {
-  skip_cert_verification = false
+  skip_cert_verification = true
 }